Massive Refactoring 3
This should hopefully be one of the last major refactoring commits for the GUI project. This refactoring aims to replace the existing chart management code and replace it with a much more flexible module system. The GUI is split into 9 parts; each containing its respective container for a module (expressed as a WPF Border in the code). Using the abstract Module class, methods such as Initialize and Update are responsible for placing a visualization on one of the containers, as well as dealing with its live data update. As a huge advantage over the previous system, modules themselves do not have the burden of fetching where the information is coming from. The ChartManager class automatically pulls each module's respective data and places it into its varMap, where the module can then freely use or modify the data as needed. Examples of this can be seen in the LineChartModule and OdometryModule, both of which have been migrated to the new abstraction based method.

As mentioned before in both the commits and commented code, a module can receive two different types of data: a configuration JSON or a data JSON. When the ChartManager has received a configuration header from the named pipe, based on its provided "chart-type" JSON key a module will be created. Initialize is soon after called using the configuration data provided in the form of the dictionary configMap.

For the GUI visualization, several changes have been made to ease the creation of modules. Rather than OxyPlot PlotViews (which happen to be restricted to only charts) being hardcoded into each container, the XAML data is now fully generated programmatically by the Initialize method. This allowed way more customization than previously achievable, as visualizations like the OdometryModule were now possible.

However, there were a few issues in having to manually write the C# code for primarily text elements, as each had to be assigned several different attributes, including text font, size, family, grid settings, and alignment configurations. Instead of doing all of this tedious work, I decided to create a ModuleTextBuilder, where several of these originally verbose methods were drastically simplified and could be easily chained together. I also noticed several text elements were repetitive in their purpose, whether it be serving as a title, subtitle, or paragraph text. To speed up the development of these, I further added a ModuleTextFactory, where predefined templates as mentioned before could be created in one line. Examples of this being used in production are seen in the OdometryModule.