🐛 Fix C#-CLI Pipe randomly breaking
I'm going to be honest, I have no idea why this change was able to fix the bug. Of all things that could possibly affect the named pipe, why is the references having the private property the issue? The program ran completely fine and even created the named pipe server, but would break as soon as data was sent through. Why?

Other minor changes:
- Moved the application dispatcher in ChartManager to encapsulate the entire foreach loop in the module updating to prevent redundancy
- Removed old odometry debugging WriteLine