✨ Theme Change & Navigational Improvements
Addressing #5, chart modules now have the ability to be synchronized. For now, this is achieved by:
1. Select the chart module you want all of the other modules to sync with. Any action, whether it be pan, zoom, scroll, or click, will select the module. Unfortunately there isn't a visual queue of this happening, and likely needs to be addressed in the future.
2. Press SPACE, signaling all of the modules to synchronize and invalidate their charts (refresh).

After a few suggestions by @thyagarajaa, I ended up redesigning the primary colors used in the UI to be a lot darker, similarly to Grafana. Some of the shadows, borders, and margins have also been changed to accompany this theme switch.

Other Misc Changes:
- Added the IconPacks nuget package for adding icon buttons later on